requirements.txt
streamlit==1.28.0
neo4j==5.14.0
pandas==2.1.3
plotly==5.17.0
python-dotenv==1.0.0
groq>=0.9.0